+++ launcher/src/com/genir/aitweaks/launcher/hullmods/SearchAndDestroy.kt
@@ -2,6 +2,7 @@ package com.genir.aitweaks.launcher.hullmods
 
 import com.fs.starfarer.api.combat.BaseHullMod
 import com.fs.starfarer.api.combat.ShipAPI
+import lunalib.lunaSettings.LunaSettings
 
 class SearchAndDestroy : BaseHullMod() {
     override fun showInRefitScreenModPickerFor(ship: ShipAPI): Boolean {
@@ -9,14 +10,18 @@ class SearchAndDestroy : BaseHullMod() {
     }
 
     override fun isApplicableToShip(ship: ShipAPI): Boolean {
-        return !ship.variant.hasHullMod("aitweaks_custom_ship_ai")
+        return !fleetwideEanbled()
     }
 
     override fun getUnapplicableReason(ship: ShipAPI): String? {
         return when {
-            ship.variant.hasHullMod("aitweaks_custom_ship_ai") -> "Ships with Custom AI always default to Search And Destroy."
+            fleetwideEanbled() -> "Fleetwide Search & Destroy is selected in LunaLib settings. All ship default to Search & Destroy assignment."
 
             else -> null
         }
+    }
+
+    private fun fleetwideEanbled(): Boolean {
+        return LunaSettings.getBoolean("aitweaks", "aitweaks_fleetwide_search_destroy") == true
     }
 }