+++ changelog.md
@@ -4,6 +4,7 @@
 - Removed mod_info.json link to ASM library to avoid conflicts with other mods that depend on different ASM versions. The library is now loaded with a custom loader. Suggested by andylizi.
 - Fixed a ClassCastException crash when calculating threat from a missile weapons firing non-missile projectiles. Reported by carbineaction and Goldendragonfinn.
 - Custom AI: Ships no longer rotate to aim weapons with DO_NOT_AIM hint, like the Voltaic Discharge. Reported by Seekers.
+- Added LunaLib setting for fleetwide Search & Destroy assignment. Suggested by lchronosl.
 
 Version v2.2.1